Some fixes for execution.

+++ src/main/java/org/spiderplan/representation/types/TypeManager.java
@@ -135,7 +135,7 @@ public class TypeManager {
 					
 					if ( this.typeNames.contains(typeName) ) {
 						Type tKnown = types.get(typeName);
-						if ( tKnown instanceof IntegerType ) {
+						if ( tKnown instanceof FloatType ) {
 							t = (FloatType)tKnown;
 						} else {
 							throw new IllegalStateException("Name " + typeName + " from " + dC + " also used for " + tKnown.getClass().getSimpleName());
